Move SubscriptionCard to separate file
